requirements.txt
matplotlib==3.10.1
numpy==1.24.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
torch==2.6.0+cu118
torchvision==0.21.0+cu118
transformers==4.49.0
